[SSD-63] refactor: css modeule.css로 수정
