+++ frontend/src/pages/signup/SignUp.module.css
@@ -1,0 +1,108 @@
+.signMainBox {
+  background-color: #f6f1e7;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+}
+
+.bigbox {
+  margin: 20px;
+}
+
+.signMainBox > div {
+  margin: 20px;
+}
+
+.fieldsBox {
+  width: 80%;
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.contentBox {
+  width: 100%;
+}
+
+.fieldsInput {
+  text-align: center;
+  width: 70%;
+  padding: 6px;
+  margin: 5px;
+}
+
+.pwfont {
+  color: red;
+  font-size: 10px;
+}
+
+.emailCheckBox {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.emailCheckBut {
+  width: 80px; /* 버튼의 가로 길이를 고정 */
+  height: 25px; /* 버튼 높이를 조정 */
+  font-size: 12px; /* 글자 크기 줄이기 */
+  background-color: #f0842c;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.borderRadius {
+  border-radius: 4px;
+  border: none;
+}
+
+.checkboxText {
+  font-size: 7px;
+}
+
+.checkboxlabel {
+  vertical-align: 4px;
+  cursor: pointer;
+}
+
+.checkBoxDiv > div {
+  font-size: 10px;
+}
+.isPrivacyPolicyAcceptedText .isTermsAcceptedTxt {
+  font-size: 10px;
+}
+
+.checkBoxDiv {
+  width: 62%;
+  display: flex; /* flexbox 사용 */
+  flex-direction: column; /* 세로 방향으로 정렬 */
+  gap: 10px; /* 각 체크박스 사이의 간격 */
+}
+
+.isPrivacyPolicyAccepted,
+.isTermsAccepted {
+  display: flex; /* flexbox 사용 */
+  align-items: center; /* 수직 정렬 */
+  justify-content: space-between; /* 좌우 간격을 동일하게 */
+}
+
+.signUpBut {
+  width: 80%;
+  height: 40px;
+  border-radius: 10px;
+  border: none;
+  background-color: #f0842c;
+  font-weight: bold;
+  font-size: 16px;
+  cursor: pointer;
+}
+
+/* 전체 요소에 드래그 방지 적용 */
+* {
+  user-select: none;
+  -webkit-user-select: none; /* Safari */
+  -moz-user-select: none; /* Firefox */
+  -ms-user-select: none; /* Internet Explorer */
+}
